requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.1.1; python_version >= '3.6'
anyio==3.6.1; python_full_version >= '3.6.2'
asgiref==3.5.2; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
billiard==3.6.4.0
black==22.8.0
celery[redis]==5.2.7
certifi==2022.9.24; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1; python_version >= '3.6'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3; python_version >= '3.7'
cryptography==38.0.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dj-rest-auth==2.2.5
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-redis==5.2.0
django==4.1.2
djangorestframework-simplejwt==5.2.1
djangorestframework==3.14.0; python_version >= '3.6'
flower==1.2.0
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
httpcore==0.15.0; python_version >= '3.7'
httpx==0.23.0
humanize==4.4.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
kombu==5.2.4; python_version >= '3.7'
lockfile==0.12.2
mypy-extensions==0.4.3
oauthlib==3.2.1; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pathspec==0.10.1; python_version >= '3.7'
platformdirs==2.5.2; python_version >= '3.7'
prometheus-client==0.14.1; python_version >= '3.6'
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
psycopg2==2.9.3
pycparser==2.21
pyjwt==2.5.0; python_version >= '3.7'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python3-openid==3.2.0
pytz==2022.4
redis==4.3.4; python_version >= '3.6'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1; python_version >= '3.7' and python_full_version < '4.0.0'
rfc3986[idna2008]==1.5.0
sentry-sdk==1.9.10
setuptools==65.4.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlparse==0.4.3; python_version >= '3.5'
tomli==2.0.1; python_full_version < '3.11.0a7'
tornado==6.2; python_version >= '3.7'
types-cryptography==3.3.23
typing-extensions==4.3.0; python_version < '3.10'
urllib3==1.26.12; python_version >= '3.6'
uvicorn==0.18.3
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
